ACT-ml/requirements.txt
openai
flask
flask_cors
python-dotenv
faiss-cpu
tqdm
numpy